// src/api/rol/errors/error.messages.ts
export enum RolMessagesError {
  ROL_ALREADY_EXISTS = 'Not possible run operation, rol already exist',
  ROL_IS_NOT_ACTIVE = 'Rol disabled',
  ROL_NOT_FOUND = 'Rol not found',
  ROL_NOT_SAVED = 'Ups... Hubo un error al crear o editar el rol. Por favor, inténtelo de nuevo más tarde',
  ROL_NOT_DELETED = 'Ups... Hubo un error al borrar el rol. Por favor, inténtelo de nuevo más tarde',
  INTERNAL_SERVER_ERROR = 'Internal server error. Contact with the administrator',
  ROL_ERROR = 'Oops... There were problems creating or editing the rol. Please try again later',
}